+++ modules/Scheduling/includes/Courses.fnc.php
@@ -555,3 +555,60 @@ function CourseDeleteSQL( $course_id )
 
 	return $delete_sql;
 }
+
+/**
+ * Automatically update schedules marking period.
+ *
+ * On the condition scheduled (old) marking period is of greater type
+ * than the new course marking period.
+ * For example: FY to SEM.
+ *
+ * @since 11.1 Move local function from Courses.php & rename to CoursePeriodUpdateMP()
+ *
+ * @param  string $cp_id Course Period ID.
+ * @param  string $mp_id Marking Period ID.
+ * @return int    Number of schedules updated.
+ */
+function CoursePeriodUpdateMP( $cp_id, $mp_id )
+{
+	global $db_connection,
+		$DatabaseType;
+
+	// Get CP MP.
+	$cp_mp = GetMP( $mp_id, 'MP' );
+
+	if ( ! $cp_id
+		|| ! $mp_id
+		|| ! $cp_mp )
+	{
+		return 0;
+	}
+
+	if ( $cp_mp === 'FY' )
+	{
+		// CP MP is Full Year, no need to update.
+		return 0;
+	}
+
+	if ( $cp_mp !== 'SEM'
+		&& $cp_mp !== 'QTR' )
+	{
+		// CP MP is not a Semester neither a Quarter...!
+		return 0;
+	}
+
+	$schedule_mp_in = ( $cp_mp === 'QTR' ? "'FY','SEM'" : "'FY'" );
+
+	// Update Schedules for CP where MP is of greater type
+	// than the new course period marking period.
+	$update = DBQuery( "UPDATE schedule SET
+		MP='" . $cp_mp . "',
+		MARKING_PERIOD_ID='" . (int) $mp_id . "'
+		WHERE COURSE_PERIOD_ID='" . (int) $cp_id . "'
+		AND MP IN (" . $schedule_mp_in . ")" );
+
+	// Return number of updated schedules.
+	return $DatabaseType === 'mysql' ?
+		mysqli_affected_rows( $db_connection ) :
+		pg_affected_rows( $update );
+}